requirement.txt
numpy==1.18.5
opencv-contrib-python==4.1.2.30
opencv-python==4.1.2.30
Pillow==7.2.0
